Code/requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
sentence-transformers
scikit-learn
plotly
matplotlib
seaborn
streamlit-lottie
